# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-oriented Go Web API scaffold. Application error codes: 0 success; 10000-19999 common/input; 20000-29999 authentication/authorization; 30000-39999 business; 50000-59999 infrastructure.",
        "title": "Go API Template",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/audit/records/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Record an immutable business audit entry",
                "parameters": [
                    {
                        "description": "Audit record",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.RecordAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.AuditRecordResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/audit/records/export": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Export a bounded audit query as CSV",
                "parameters": [
                    {
                        "description": "Audit export filter",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ExportAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportAuditResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/audit/records/get": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Get an audit record",
                "parameters": [
                    {
                        "description": "Audit ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.GetAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.AuditRecordResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/audit/records/query": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Query audit records",
                "parameters": [
                    {
                        "description": "Audit filter",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.QueryAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.AuditPageResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Return the authenticated subject",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.MeResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Code 20001: unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httptransport.AuditPageResponseBody": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httptransport.AuditRecordResponseBody"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "httptransport.AuditRecordResponseBody": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "after_summary": {
                    "type": "object"
                },
                "application_id": {
                    "type": "string"
                },
                "before_summary": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "source_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httptransport.ExportAuditRequest": {
            "type": "object",
            "required": [
                "tenant_id"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "application_id": {
                    "type": "string"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "keyword": {
                    "type": "string"
                },
                "max_records": {
                    "type": "integer"
                },
                "occurred_from": {
                    "type": "string"
                },
                "occurred_to": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "source_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ExportAuditResponseBody": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "record_count": {
                    "type": "integer"
                }
            }
        },
        "httptransport.GetAuditRequest": {
            "type": "object",
            "required": [
                "id",
                "tenant_id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.MeResponseBody": {
            "type": "object",
            "properties": {
                "subject": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.QueryAuditRequest": {
            "type": "object",
            "required": [
                "tenant_id"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "application_id": {
                    "type": "string"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "keyword": {
                    "type": "string"
                },
                "occurred_from": {
                    "type": "string"
                },
                "occurred_to": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "source_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.RecordAuditRequest": {
            "type": "object",
            "required": [
                "action",
                "resource_id",
                "resource_type",
                "source_service",
                "tenant_id"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "after_summary": {
                    "type": "object"
                },
                "application_id": {
                    "type": "string"
                },
                "before_summary": {
                    "type": "object"
                },
                "id": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "source_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.Response": {
            "type": "object",
            "properties": {
                "body": {},
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter \"Bearer {JWT}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PSK": {
            "description": "Enter \"PSK {shared-key}\" for routes configured with PSK authentication.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}